Introduce Dagger for DI
