+++ gradle.properties
@@ -6,3 +6,4 @@ shadow_version=5.1.0
 versions_version=0.26.0
 ktorm_version=2.5
 postgresql_version=42.2.8
+dagger_version=2.24
