+++ src/com/tylerkindy/chooseit/Main.kt
@@ -1,5 +1,6 @@
 package com.tylerkindy.chooseit
 
+import com.tylerkindy.chooseit.di.DaggerAppComponent
 import com.tylerkindy.chooseit.model.Room
 import com.tylerkindy.chooseit.model.Rooms
 import io.ktor.application.call
@@ -16,6 +17,7 @@ import io.ktor.response.respond
 import io.ktor.routing.get
 import io.ktor.routing.post
 import io.ktor.routing.routing
+import io.ktor.server.engine.ApplicationEngine
 import io.ktor.server.engine.embeddedServer
 import io.ktor.server.netty.Netty
 import me.liuwj.ktorm.database.Database
@@ -33,62 +35,7 @@ fun main() {
         driver = "org.postgresql.Driver"
     )
 
-    val server = embeddedServer(Netty, port = System.getProperty("server.port").toInt()) {
-        install(ContentNegotiation) {
-            gson { }
-        }
-        install(CallLogging)
-        install(Locations)
-        install(CORS) {
-            anyHost()
-        }
-
-        routing {
-            get("/rooms") {
-                val roomIds = Rooms.select(Rooms.id).map { it[Rooms.id]!! }
-
-                call.respond(
-                    mapOf(
-                        "rooms" to roomIds
-                    )
-                )
-            }
-
-            post("/new-room") {
-                val uuid = UUID.randomUUID()
-                val bb = ByteBuffer.wrap(Array<Byte>(16) { 0 }.toByteArray())
-                bb.putLong(uuid.mostSignificantBits)
-                bb.putLong(uuid.leastSignificantBits)
-                val id = Base64.getUrlEncoder().encodeToString(bb.array())!!
-
-                Rooms.insert {
-                    it.id to id
-                }
-
-                call.respond(id)
-            }
-
-            get<RoomRoute.Status> { params ->
-                val room = getRoom(params.room.id)
-                val value = when (room.flip) {
-                    true -> "Heads"
-                    false -> "Tails"
-                    null -> "Not yet flipped"
-                }
-
-                call.respond(value)
-            }
-
-            post<RoomRoute.Flip> { params ->
-                val room = getRoom(params.room.id)
-
-                room.flip = Random.nextBoolean()
-                room.flushChanges()
-                call.respond("OK")
-            }
-        }
-    }
-
+    val server: ApplicationEngine = DaggerAppComponent.create().server()
     server.start(wait = true)
 }
