+++ src/com/tylerkindy/chooseit/di/AppComponent.kt
@@ -1,0 +1,11 @@
+package com.tylerkindy.chooseit.di
+
+import dagger.Component
+import io.ktor.server.engine.ApplicationEngine
+
+@Component(modules = [
+    MainModule::class
+])
+interface AppComponent {
+    fun server(): ApplicationEngine
+}
